XWIKI-9006: The schema does not seems to be deleted when deleting a wiki using PostgreSQL in schema mode * Check that we are in schema mode (thanks to Thomas!)

+++ xwiki-platform-core/xwiki-platform-oldcore/src/main/java/com/xpn/xwiki/store/XWikiHibernateStore.java
@@ -410,7 +410,11 @@ public class XWikiHibernateStore extends XWikiHibernateBaseStore implements XWik
             } else if (DatabaseProduct.MYSQL == databaseProduct) {
                 stmt.execute("DROP DATABASE " + escapedSchema);
             } else if (DatabaseProduct.POSTGRESQL == databaseProduct) {
-                stmt.execute("DROP SCHEMA " + escapedSchema + " CASCADE");
+                if (isInSchemaMode()) {
+                    stmt.execute("DROP SCHEMA " + escapedSchema + " CASCADE");
+                } else {
+                    LOGGER.warn("Subwiki deletion not yet supported in Database mode for PostgreSQL");
+                }
             }
 
             endTransaction(context, true);
